BarChart component.
