+++ components/metrics/EventsChart.module.css
@@ -1,0 +1,3 @@
+.chart {
+  display: flex;
+}
